tutorials/rockset-mongo-integration/requirements.txt
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
requests-toolbelt==0.9.1
rockset==0.7.6b20200423
timeloop==1.0.2
